Added Shield functionality, enemy projectile damage, mineable environment object functionality
Improved the sword and shield functionality. Added the enemy projectile functionality for dealing damage to the player. Added mineable environment objects with functionality (stone for now but there is a template made for more). If hitting the object with a sword then the material will spawn. Added spawning the environment object onto the terrain.